# chore(epar): example of improved performance