--- a/Cargo.toml
+++ b/Cargo.toml
@@ -34,6 +34,7 @@ fuel-types = { version = "0.60.0", path = "fuel-types", default-features = false
 fuel-vm = { version = "0.60.0", path = "fuel-vm", default-features = false }
 bitflags = "2"
 bincode = { version = "1.3", default-features = false }
+bn = { package = "substrate-bn", version = "0.6", default-features = false }
 criterion = "0.5.0"
 
 [profile.web-release] # For minimal wasm binaries, use this profile
@@ -43,3 +44,6 @@ codegen-units = 1
 lto = true
 strip = true
 panic = "abort"
+
+[patch.crates-io]
+bn = { git = "https://github.com/rymnc/bn", package = "substrate-bn", rev = "ae78ed4" }