--- a/fuel-vm/Cargo.toml
+++ b/fuel-vm/Cargo.toml
@@ -16,12 +16,17 @@ name = "execution"
 harness = false
 required-features = ["std"]
 
+[[bench]]
+name = "epar"
+harness = false
+required-features = ["std"]
+
 [dependencies]
 anyhow = { version = "1.0", optional = true }
 async-trait = "0.1"
 backtrace = { version = "0.3", optional = true } # requires debug symbols to work
 bitflags = { workspace = true }
-bn = { package = "substrate-bn", version = "0.6", default-features = false }
+bn = { workspace = true }
 derive_more = { version = "0.99", default-features = false, features = [
     "display",
 ] }