> [!WARNING]
> this probably shouldn't be merged in, but this PR serves as a general idea on how to squeeze out 10% improvement for the elliptic curve pairing opcode (`alt_bn_128`). There are a few more optimizations i haven't applied yet, like simd or loop unrolling. i also see that we're doing unnecessary conversions from the affine repr to projective and then back to affine within the pairing fn imported from the `substrate-bn` lib. perhaps we should consider using another lib, but its fine for now ig. ecop needs to be optimized.

[Link to related issue(s) here, if any]
- upstream patch: https://github.com/rymnc/bn/commit/ae78ed4e11e59f7ede0003ae61a849be8553882f


## Description

Small trivial fixes to the upstream lib via patch -
- fixed vec allocation (causes upfront overhead for the full allocation)
- reused affine conversions
- some early returns

you can also notice stability/lower outlier count due to the upfront preallocation. the miller loop is still unoptimized, and that will be the main source of perf gain later.

### Before

```
epar/test_vector        time:   [550.58 ms 552.37 ms 554.37 ms]
Found 10 outliers among 100 measurements (10.00%)
  5 (5.00%) high mild
  5 (5.00%) high severe
```

### After

```
epar/test_vector        time:   [492.48 ms 494.92 ms 497.21 ms]
                        change: [-10.5525% -10.4000% -10.3108%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```

## Checklist
- [ ] Breaking changes are clearly marked as such in the PR description and changelog
- [ ] New behavior is reflected in tests
- [ ] If performance characteristic of an instruction change, update gas costs as well or make a follow-up PR for that
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/) matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [ ] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+) (for out-of-organization contributors, the person merging the PR will do this)
- [ ] Someone else?